requirements.txt
transformers
# torch
python-dotenv
bitsandbytes
accelerate
sentencepiece
protobuf
datasets
# pip3 install torch --index-url https://download.pytorch.org/whl/cu124